requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.17.0
alembic==1.17.0
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.0
asyncpg==0.30.0
asyncpraw==7.8.1
asyncprawcore==2.4.0
attrs==25.4.0
bcrypt==5.0.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.119.0
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
idna==3.11
iniconfig==2.3.0
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.0
mypy==1.18.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pluggy==1.6.0
propcache==0.4.1
psycopg2-binary==2.9.11
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
python-dotenv==1.1.1
redis==7.0.1
requests==2.32.5
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
update-checker==0.18.0
urllib3==2.5.0
uvicorn==0.38.0
vaderSentiment==3.3.2
websocket-client==1.9.0
yarl==1.22.0
zxcvbn==4.5.0
